+++ src/sqlfluff/core/templaters/builtins/__init__.py
@@ -1,0 +1,1 @@
+"""Module for defining context builtins."""
